backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
openai==1.41.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
psycopg2-binary==2.9.9
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
python-dotenv==1.0.0
sentence-transformers==3.0.1
tiktoken==0.7.0
Werkzeug==3.0.3
gunicorn==20.1.0
SQLAlchemy==2.0.30
cloud-sql-python-connector==1.12.0
pg8000==1.31.2
